Moved files around as there were duplicate folder names
