--- backend-springboot/src/main/java/com/cscd488seniorproject/syllabussyncproject/controller/dto/CanvasSubscribeRequest.java
@@ -1,6 +1,0 @@
-package com.cscd488seniorproject.syllabussyncproject.controller.dto;
-
-public record CanvasSubscribeRequest(String icsUrl) {
-
-}
-
