+++ backend-springboot/src/main/java/com/cscd488seniorproject/syllabussyncproject/dto/CanvasSubscribeRequest.java
@@ -1,0 +1,6 @@
+package com.cscd488seniorproject.syllabussyncproject.dto;
+
+public record CanvasSubscribeRequest(String icsUrl) {
+
+}
+
